requirements.txt
openai==1.64.0
twitter-api-client==0.10.11
python-dotenv==1.0.0